fixbug023

+++ src/new_embedder.py
@@ -634,7 +634,7 @@ def get_linear_tgt_model(args, root, sample_shape, num_classes, loss,lora_rank =
             src_ys.append(y_.detach().cpu())
             src_feats.append(out.detach().cpu())
     src_feats = torch.cat(src_feats, 0)
-    src_ys = torch.cat(src_ys, 0).long()
+    src_ys = torch.cat(src_ys, 0)
     src_train_dataset = torch.utils.data.TensorDataset(src_feats, src_ys)        
     del src_model, src_train_loader
